% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilaritiesMethods.R
\name{as.treesimilarities.list}
\alias{as.treesimilarities.list}
\title{A function to coerce objects to a treesimilarities object.}
\usage{
\method{as.treesimilarities}{list}(obj, ...)
}
\arguments{
\item{obj}{List with as first element the similarity matrix, and as second the forest on which it was calculated}

\item{...}{Additional arguments}
}
\description{
A function to coerce objects to treesimilarities objects. It takes a list with a similaritymatrix as its first element
and a forest object as its second, as arguments.
This list is then coerced to a treesimilarities object (with attributes similarity matrix and forest).
}
