% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMethods.R
\name{clusters.clusterforest}
\alias{clusters.clusterforest}
\title{Get the cluster assignments for a solution of a clusterforest object}
\usage{
\method{clusters}{clusterforest}(clusterforest, solution = 1)
}
\arguments{
\item{clusterforest}{The clusterforest object}

\item{solution}{The solution}
}
\description{
A function to get the cluster assignments for a given solution of a clusterforest object.
}
