% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMethods.R
\name{clusters}
\alias{clusters}
\title{Get the cluster assignments for a solution of a clusterforest object}
\usage{
clusters(clusterforest, solution)
}
\arguments{
\item{clusterforest}{A clusterforest object}

\item{solution}{The solution for which cluster assignments should be returned. Default = 1}
}
\description{
A function to get the cluster assignments for a given solution of a clusterforest object.
}
