% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsForest.R
\name{plot.forest}
\alias{plot.forest}
\title{Plot a forest object}
\usage{
\method{plot}{forest}(x, ..., treetoplot = 1)
}
\arguments{
\item{x}{A clusterforest object}

\item{...}{Additional arguments}

\item{treetoplot}{The tree from the forest to be plotted. Default=1}
}
\description{
A function to plot a tree from a forest.
}
