% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilaritiesMethods.R
\name{plot.treesimilarities}
\alias{plot.treesimilarities}
\title{Plot a treesimilarities object}
\usage{
\method{plot}{treesimilarities}(x, ...)
}
\arguments{
\item{x}{A treesimilarities object}

\item{...}{Additional arguments}
}
\description{
A function to plot a treesimilarities object.
}
