% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsForest.R
\name{print.forest}
\alias{print.forest}
\title{Print a forest object}
\usage{
\method{print}{forest}(x, ..., treetoprint = NULL)
}
\arguments{
\item{x}{A clusterforest object}

\item{...}{Additional arguments}

\item{treetoprint}{The tree to print. If NULL then all trees are printed. Default=NULL.}
}
\description{
A function to print a forest, or a tree from a forest.
}
