% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilaritiesMethods.R
\name{print.treesimilarities}
\alias{print.treesimilarities}
\title{print a treesimilarities object}
\usage{
\method{print}{treesimilarities}(x, ...)
}
\arguments{
\item{x}{A treesimilarities object}

\item{...}{Additional arguments}
}
\description{
A function to print a treesimilarities object.
}
