% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methodsForest.R
\name{summary.forest}
\alias{summary.forest}
\title{Summarize a forest object}
\usage{
\method{summary}{forest}(object, ...)
}
\arguments{
\item{object}{A forest object}

\item{...}{Additional arguments}
}
\description{
A function to summarize a forest.
}
