% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimilaritiesMethods.R
\name{summary.treesimilarities}
\alias{summary.treesimilarities}
\title{Summarize a treesimilarities object}
\usage{
\method{summary}{treesimilarities}(object, ...)
}
\arguments{
\item{object}{A treesimilarities object}

\item{...}{Additional arguments}
}
\description{
A function to summarize a treesimilarities object.
}
