% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMethods.R
\name{treesimilarities.clusterforest}
\alias{treesimilarities.clusterforest}
\title{Get the similarity matrix that wast used to create a clusterforest object}
\usage{
treesimilarities.clusterforest(x, ...)
}
\arguments{
\item{x}{A clusterforest object}

\item{...}{Additional arguments}
}
\description{
A function to get the similarity matrix used to obtain a clusterforest object.
}
