% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMethods.R
\name{treesimilarities.default}
\alias{treesimilarities.default}
\title{Get the similarity matrix that wast used to create a clusterforest object}
\usage{
\method{treesimilarities}{default}(clusterforest)
}
\arguments{
\item{clusterforest}{A clusterforest object}
}
\description{
A function to get the similarity matrix used to obtain a clusterforest object.
}
