% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treesourceClusterMethod.R
\name{treesource.clusterforest}
\alias{treesource.clusterforest}
\title{Mapping the tree clustering solution to a known source of variation underlying the forest}
\usage{
\method{treesource}{clusterforest}(clusterforest, solution)
}
\arguments{
\item{clusterforest}{The clusterforest object}

\item{solution}{The solution}
}
\description{
A function that can be used to get insight into a clusterforest solution, in the case that there is a known
source of variation underlying the forest.
It visualizes the number of trees from each source that belong to each cluster.
}
