% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMethods.R
\name{clusters.clusterforest}
\alias{clusters.clusterforest}
\title{Get the cluster assignments for a solution of a clusterforest object}
\usage{
clusters.clusterforest(x, ..., solution = 1)
}
\arguments{
\item{x}{A clusterforest object}

\item{...}{Additional arguments}

\item{solution}{The solution for which cluster assignments should be returned. Default = 1}
}
\description{
A function to get the cluster assignments for a given solution of a clusterforest object.
}
