% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMethods.R
\name{medoidtrees.clusterforest}
\alias{medoidtrees.clusterforest}
\title{Get the medoid trees for a solution of a clusterforest object}
\usage{
medoidtrees.clusterforest(x, ..., solution = 1)
}
\arguments{
\item{x}{A clusterforest object}

\item{...}{Additional arguments}

\item{solution}{The solution for which medoid trees should be returned. Default = 1}
}
\description{
A function to get the medoid trees for a given solution of a clusterforest object.
}
