% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMethods.R
\name{print.clusterforest}
\alias{print.clusterforest}
\title{Print a clusterforest object}
\usage{
\method{print}{clusterforest}(x, ..., solution = 1)
}
\arguments{
\item{x}{A clusterforest object}

\item{...}{Additional arguments}

\item{solution}{The solution to print the medoid trees from. Default = NULL}
}
\description{
A function that can be used to print a clusterforest object.
}
