% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterMethods.R
\name{summary.clusterforest}
\alias{summary.clusterforest}
\title{Summarize a clusterforest object}
\usage{
\method{summary}{clusterforest}(object, ...)
}
\arguments{
\item{object}{A clusterforest object}

\item{...}{Additional arguments}
}
\description{
A function to summarize a clusterforest object.
}
